Optimize Map Generation Algorithm
This commit makes dire changes to the map algorithm. The original time to generate the map took 11 minutes.

After implementing a more efficient priority queue (fully generated by ChatGPT), this time was cut down to 3 minutes

After implementing A*, the time to generate the map remained unchanged, meaning the bottleneck was not with the pathfinding algorithm itself.

After changing the GetPathLength algorithm, the time to generate the map went from 3 minutes to 10 seconds. This is largely due to the super inefficient calls to pathfind each time you want to get the path length. Instead, you can calculate the Manhattan distance, which provides a correct path length regardless of the path itself.

Further optimizations may be made in the future. The current bottleneck is no longer with the pathfinding, nor the edge connectors. The bottleneck is now with the chamber generation algorithm. This takes 6 seconds to generate, according to the debugged time. So, for future optimizations, direct your attention to the chamber generation algorithm.