Add Hallway/Chamber Prefabs || Determine Correct Hallway Connector to Place
This commit finalizes the content for setting up the map generation process. This commit adds larger chamber models to be placed on the map. Same with the hallway model. There are 6 different hallway models to choose from based on the previous and next hallway connector.

The last thing that will be done is some code refactoring for professionalism. I neglected the code cleanup and have many useless variable. I will create simple getters for chamber properties that we may use in later programming.