Add Rooms
This commit adds the rooms for 1x4, 2x2, 2x3, and 3x2 rooms on behalf of Aiden. The pipeline for adding the rooms will need to be changed. The rooms need to be made in Unity instead of Blender. We will discuss the changes that need to be made in our next daily meeting.