Update 2x3/3x2 Rooms
This commit finishes the updates for the rooms to work with prefab assets. From now on, the rooms will be made in Unity by dragging & dropping the assets into the room. That way, we may skip the process of changing the assets from the Blender asset into the Unity prefab asset.