Fix Seed RNG Bug
This commit fixes the RNG bug, where the player would input the seed & the map would come out differently.

The problem: I referenced the textmeshprougui component where the player was inputting their game seed. the issue with this is the component value was stored statically, so the player input text was always "", which is not an integer (causing the seed to be set to a random number).

The Solution: Change the reference to the input field component, not the textmeshprougui. Get its text value and that will return the player's input. For future reference, textmeshprougui's value is read statically, not dynamically. changing its value does not change reference values. Be cautious in the future doing this.