Add Initialization Logic for Editor Runtime
This commit adds an editor script that handles singletons and managers that scenes need. When playing from the Unity editor, the script, PlayFromInitScene.cs, will load the dev into the Initialization scene. When exiting play mode (entering edit mode), the script will reopen the scene that the dev started playing from.

The commit also creates the system for the initialization scene. It holds a game object with a script to initialize singletons/managers. Given the singletons/managers, it will create each game object during awake. During start, it will determine if the game should load into the main menu scene (which will always be the case for any build of the game), or if it should load into a previous scene opened by the developer.

TL;DR, this commit ensures that singletons/managers are never null when playing from the editor. Just add singletons to the initObj prefab.