Add Dynamic Mouse Locking/Unlocking Based on InputManager || Develop System for Initial Input Scheme on Scene Loading
This commit adds the long-awaited feature of locking and hiding the user's mouse when the game is in session. When the game is run, the mouse will be hidden from the user and its position will be locked. The only time this is not true is when the scene startup is set to enable the mouse. More about that below.

I've added some additions for handling the input scheme when initializing the scene. When the scene is loaded, the input manager will hear the event call and handle whether the mouse will be locked or unlocked. This is done through the caching of scene asset names in scriptable objects. You can access the names of scenes through the scene asset, but this will not work when building the game. Thus, using an SO as a middleman fixes the problem.