Implement Object Culling
This commit introduces chamber and edge culling. Now, chambers will only be rendered if and only if it is the current room the player is in or it is a room adjacent to the current room. For the edges, the edges from the current room to adjacent rooms are rendered, and then the edges connected to that chamber are rendered.

There is a slight bug where some edges will not properly cull or render, even through the current chamber connects between them. I will have to try a different approach to solve the edge culling & rendering problem later.