Update Edge Culling/Rendering
This commit fixes the bug with the edges being culled/rendered. My original idea overcomplicated the problem and let to some confusing code.

My revised change was to have the edge check if it should be culled. If either chamber next to it is rendered, then it should render. If neither are rendered, then it should be culled.

Considerations for in-game performance is still in question. Theoretically, the culling process can brick performance for a split second, causing some stutters when entering a chamber. Future optimizations should aim to find an asynchronous solution to the rendering/culling.