Add the Map
This commit adds in the overall map to the world. You can see the map in the world generation map test scene. Next up is creating a minimap.

We will need someone to make the UI artwork for chambers, hallways, and other sprites that will be shown on the map display.