Simple Inventory Setup || Add Placement Logic
This commit adds the logic for adding an item to the inventory and defining the inventory. Note that the inventory grid positions start at the top-left most corner where the x-component of the grid position increases from left to right, and the z-component increases from top to bottom.

A simple test was done to add an item to the inventory. What is needed now is the add the logic to stop the addition of the item at the slot if it interferes with another item in the inventory, and the logic for picking up the item needs to be added. Additionally, the logic for moving between two inventories need to be defined.