Add Player Pickup/Dropoff Logic
This commit adds the logic for the player picking up and dropping off an item into an inventory. To do this, a player cursor inventory singleton was created. It holds a single item of any structure. When the player doesn't have an item, it will pick up any item from whichever position in the inventory the player clicked. When they have picked up an item, the inventory will attempt to place the item at the corresponding position. If is succeeds, it will go through. Otherwise, the player continues to hold that item.

TODO: Add another inventory for the player backpack (the current one is named the backpack, but it really is the augment inventory). Add the ability to expand the augment and backpack inventory with a simple method. Add some other item structures for testing purposes, and add the ability to open/close the inventory UI.