Add Inherited Inventory Systems || Add Inventory Toggling in Input Manager
This commit adds the ability to open and close the inventory using the input manager singleton. It also adds specific types of inventories (augment inventory and backpack inventory).

I've decided to leave the base class the same, only making some methods public/protected virtual. The reasoning for this is because not all inventories will be specialized, unless we want them to. The backpack and augmented inventory have a single use-case different from the normal inventory system, so they get to inherit. But a typical inventory does not need custom logic different from the base.

The last thing to add to the inventory system is to add the ability to pick up objects in the world and add them to the inventory accordingly. I created a method to try to automatically add an item to an inventory, so items can run an OnPickup method to be added to the inventory.