Add Item Pickup Logic
This commit adds the ability for the player to pickup items in the world and add it the the backpack.

The actual logic has the items implement the IPickup custom interface which defines a singular OnPickup method. It is up to the item class to define the logic.

Future improvements to this system should aim to make items inherit from the Item class. Then, each item will have specialized item pickup logic (maybe not all items should be put into the backpack).

Other improvements that may be added are some inventory quality-of-life improvements. Below are some additions that may be made:
->Add more methods to handle item movement.
   ->if placing an item is blocked by a singular item when manually placing, swap that item with the cursor item. Basic swap logic
   ->shift-click automatically moves items in the inventory to the destined inventory system (i.e., augments move from the backpack to the augment inventory without manually placing it)
   ->right-click to open extra options for what to do with an item (i.e., discard, drop, and other effects).
->Hovering over an item displays its name, description, and effects (if applicable)
->Replace the placeholder sprites with quality sprites. Also design a UI interface that looks neat.