Add Basic Inventory System (#29)
* Add Basic Augment Inventory System

This commit creates the inventory structure for the augment inventory. More inventories are to come, as well as the logic for adding items to an inventory, moving them, and removing them.

* Simple Inventory Setup || Add Placement Logic

This commit adds the logic for adding an item to the inventory and defining the inventory. Note that the inventory grid positions start at the top-left most corner where the x-component of the grid position increases from left to right, and the z-component increases from top to bottom.

A simple test was done to add an item to the inventory. What is needed now is the add the logic to stop the addition of the item at the slot if it interferes with another item in the inventory, and the logic for picking up the item needs to be added. Additionally, the logic for moving between two inventories need to be defined.

* Add Player Pickup/Dropoff Logic

This commit adds the logic for the player picking up and dropping off an item into an inventory. To do this, a player cursor inventory singleton was created. It holds a single item of any structure. When the player doesn't have an item, it will pick up any item from whichever position in the inventory the player clicked. When they have picked up an item, the inventory will attempt to place the item at the corresponding position. If is succeeds, it will go through. Otherwise, the player continues to hold that item.

TODO: Add another inventory for the player backpack (the current one is named the backpack, but it really is the augment inventory). Add the ability to expand the augment and backpack inventory with a simple method. Add some other item structures for testing purposes, and add the ability to open/close the inventory UI.

* Add Extra Item Structures || Add Ability to Expand Inventories

This commit adds some extra items to test the bounds of the inventory. It also adds the ability to add extra inventory slots.

TODO: Add the ability to open/close the inventory UI, make inventory system script abstract & have backpack / augment inventory inherit from it

The reason why the inventory system should become abstract is because the backpack and augment system require separate logic, yet are quite similar. This gives the inventories the ability to add their own custom logic without interfering with each other.

* Add Inherited Inventory Systems || Add Inventory Toggling in Input Manager

This commit adds the ability to open and close the inventory using the input manager singleton. It also adds specific types of inventories (augment inventory and backpack inventory).

I've decided to leave the base class the same, only making some methods public/protected virtual. The reasoning for this is because not all inventories will be specialized, unless we want them to. The backpack and augmented inventory have a single use-case different from the normal inventory system, so they get to inherit. But a typical inventory does not need custom logic different from the base.

The last thing to add to the inventory system is to add the ability to pick up objects in the world and add them to the inventory accordingly. I created a method to try to automatically add an item to an inventory, so items can run an OnPickup method to be added to the inventory.

* Add Item Pickup Logic

This commit adds the ability for the player to pickup items in the world and add it the the backpack.

The actual logic has the items implement the IPickup custom interface which defines a singular OnPickup method. It is up to the item class to define the logic.

Future improvements to this system should aim to make items inherit from the Item class. Then, each item will have specialized item pickup logic (maybe not all items should be put into the backpack).

Other improvements that may be added are some inventory quality-of-life improvements. Below are some additions that may be made:
->Add more methods to handle item movement.
   ->if placing an item is blocked by a singular item when manually placing, swap that item with the cursor item. Basic swap logic
   ->shift-click automatically moves items in the inventory to the destined inventory system (i.e., augments move from the backpack to the augment inventory without manually placing it)
   ->right-click to open extra options for what to do with an item (i.e., discard, drop, and other effects).
->Hovering over an item displays its name, description, and effects (if applicable)
->Replace the placeholder sprites with quality sprites. Also design a UI interface that looks neat.