Setup Abstract Interactable Class || Player Crossbow Fixes || Add Doors to Chambers
This commit adds an abstract interactable class (not an interface) for classes to inherit from. The player has a new script and child game object to handle the interactions with other game objects. When interacted with, the script will display a message to interact with the object. The input manager received a new binding for this new functionality and was set up properly.

The player crossbow projectile had some problems with its setup. The projectiles was made a trigger with a rigidbody, and it destroyed itself on trigger enter with anything else that wasn't a projectile. This meant the box collider triggers in the chambers interfered with the projectile, causing it to destroy itself instantly. I made the projectiles not a trigger and moved the trigger enter logic into the oncollisionenter unity event instead. Now, the projectile can properly spawn without worrying about the box triggers for chambers.

The chamber prefabs were changed to have doorways that you can interact with where appropriate. They still need to be desiged to lock and not be interacted with when the wave event is happening. For now, they are set to inactive on interaction. This can be set up properly when we get the wave manager fully functional.

The door script was made. It inherits from the Interactable class and applies its own logic for interactions when playing.

Next commits under this branch will have items designated to each chamber. However, the full setup will be restricted without the fully functional wave manager to drop the item on wave complete.