Represent Doorways by Their Position in Chamber || Make Doors Open on Both side On Door Innteract
This commit adds some grid positions to the door script to better handle their logic when one door in an edge is interacted with. When a door is interacted with, it will open, and the door on the other side will open too.

Some chamber prefabs were updated to better help with this new change. The door position needs to be assigned within the chamber prefab, both through code or a scriptable object. Reference the chamber design document to find the door positions. They are the red squares in the document.