Block Doors that Lead to Nothing
This commit will block the doorways that will lead to nothing. The chamber script gained a new dictionary of used hallway connectors. I forgot that the hallway connectors that exist on the outer bounds of the map will be destroyed, so they do not exist in the hallwayConnectors dictionary. Because of that, all doors on the edge of the map were technically interactable in the previous setup.