Update Chamber Prefab Design || Update ChamberLayoutSO
This commit updates the current chamber designs we have to follow the new structure. The new structure has the following pipeline to creating a chamber:

1) create a prefab variant of the base chamber prefab object. This is located under Assets/Map Generation/Chamber Prefabs.
2) Add your prefab variant to the Assets/Map Generation/Chamber Prefabs/Prefab Design folder under a new folder for the chamber design
3) Create the outline of the chamber using the assets under Assets/Map Generation/Chamber Prefabs/Chamber Building Block Prefabs. Create the wall outline, the doorways, and add a floor and ceiling.
4) After creating the outline, create a prefab variant of that chamber design. Create one for the construction site, underground, and hellscape chamber, or leave some blank if there isn't a prefab variant for that chamber.

The chamber layout SO script was also updated to include variants for the underground and hellscape prefab design. These do not need to be filled in - I will program the map system such that null chamber prefabs are excluded from the map generation.