Add Boss Chamber || Tidy Up Chamber Prefabs || Fix Edge Connection Bug
Found a bug where edges will connect to the same hallway connected for a chamber. The issue appeared to be because of the initial connection not checking if the chamber's hallway connector was already used. This shouldn't need to be done since the initial connection already prevented more than 1 connection, but doing the double-check appears to have fixed the bug.