Chamber designs and systems (#42)
* Small Fix for Chamber 1x4

Fixed scaffolding not within renderer

* Fix Culling Issues

* Represent Doorways by Their Position in Chamber || Make Doors Open on Both side On Door Innteract

This commit adds some grid positions to the door script to better handle their logic when one door in an edge is interacted with. When a door is interacted with, it will open, and the door on the other side will open too.

Some chamber prefabs were updated to better help with this new change. The door position needs to be assigned within the chamber prefab, both through code or a scriptable object. Reference the chamber design document to find the door positions. They are the red squares in the document.

* Block Doors that Lead to Nothing

This commit will block the doorways that will lead to nothing. The chamber script gained a new dictionary of used hallway connectors. I forgot that the hallway connectors that exist on the outer bounds of the map will be destroyed, so they do not exist in the hallwayConnectors dictionary. Because of that, all doors on the edge of the map were technically interactable in the previous setup.

* Add Door Opening Animcation

This commit adds a responsive animation to the doors when they are opened. Gone are the days where they disappear on interacting with.

* Update Chamber Prefab Design || Update ChamberLayoutSO

This commit updates the current chamber designs we have to follow the new structure. The new structure has the following pipeline to creating a chamber:

1) create a prefab variant of the base chamber prefab object. This is located under Assets/Map Generation/Chamber Prefabs.
2) Add your prefab variant to the Assets/Map Generation/Chamber Prefabs/Prefab Design folder under a new folder for the chamber design
3) Create the outline of the chamber using the assets under Assets/Map Generation/Chamber Prefabs/Chamber Building Block Prefabs. Create the wall outline, the doorways, and add a floor and ceiling.
4) After creating the outline, create a prefab variant of that chamber design. Create one for the construction site, underground, and hellscape chamber, or leave some blank if there isn't a prefab variant for that chamber.

The chamber layout SO script was also updated to include variants for the underground and hellscape prefab design. These do not need to be filled in - I will program the map system such that null chamber prefabs are excluded from the map generation.

* Add Boss Chamber || Tidy Up Chamber Prefabs || Fix Edge Connection Bug

Found a bug where edges will connect to the same hallway connected for a chamber. The issue appeared to be because of the initial connection not checking if the chamber's hallway connector was already used. This shouldn't need to be done since the initial connection already prevented more than 1 connection, but doing the double-check appears to have fixed the bug.

* Fix Sticky Walls || Increase Player Step Height

Removed friction from walls and increased the player's step height. The player's step height was set so low that the smallest bumps would block all movement. This commit changes that.

* Update Remaining Chamber Prefabs || Remove Old Chamber Prefabs

This commit updates the chamber prefabs to use the prefab variant hierarchy. You may find all of the chamber layouts on the chamber design documentation in the Teams channel.

All chamebrs have their respective chamber layout SO completed, and the base prefab for each chamber has been design.

TODO: Update the colliders in all chambers, & add construction site variants to the chambers. Also, update the existing chambers to remove the boxed feeling (Add some interior design)

---------

Co-authored-by: Jandel <127899373+Jandel456@users.noreply.github.com>