Add Health Script & IDamageable Interface || Update Construction Site Prefabs
This commit adds some core systems to the game: health and damageable interface. These components can be used on anythign that has health behaviour or needs to take damage in the world.

A base object prefab was created, and I am going through the process of updating all of our original object prefabs to follow this new system.

Additionally, this commit introduces the ProBuilder package. This will be used to cut objects into small components. Then, we may make objects destructable.