Add Object Destruction
this commit adds object destruction. Now, if you hit an object enough times, it will break and blow up into many small pieces. The pieces are cleaned up after a small delay, so no need to worry about objects flooding the scene.

The next commit will fix the colliders on the scaffold when it gets scaled to a different size. the mesh and colliders scale differently, causing invisible walls to appear.