Object assets | Update Object Prefabs to Prefab Variants (#43)
* Added New FBX Files

* Move Files, Make Objects into Prefabs

* Mine Rooms Made

* FBX Hellscape Files Added

* Added New Door Files

* Changed Hallway Sizes

* Filled Hallway Prefabs

* Added 2 FBX Enemy Files

* Add Health Script & IDamageable Interface || Update Construction Site Prefabs

This commit adds some core systems to the game: health and damageable interface. These components can be used on anythign that has health behaviour or needs to take damage in the world.

A base object prefab was created, and I am going through the process of updating all of our original object prefabs to follow this new system.

Additionally, this commit introduces the ProBuilder package. This will be used to cut objects into small components. Then, we may make objects destructable.

* Update Mining Center Prefabs

This commit updates all of the mining center prefabs to follow the prefab variant structure.

All that is left to do now is replace all the original prefab references in the chambers and use the new prefab variants

* Update Object Prefab References in Chambers & Player

This commit finalizes the changes for all references to the original prefabs and makes them reference the new prefab variants.

* Add Hellscape Prefab Variants

This commit adds prefab variants to the Hellscape fbx files

* Update Object Colliders

This commit updates many of the colliders on the object prefabs. I used primitive colliders to make up the complex shapes. Additionally, I made the rock hook object with the chain act like a physical chain thanks to the hinge joint object component. We should explore adding more chains to our levels

* Add Destroyed Construction Site Objects

This commit adds the destroyed version of the objects for the construction site. Future commits will get the remaining objects in. First, I will test the code to ensure everything works correctly when destroyed the real game objects.

* Add Object Destruction

this commit adds object destruction. Now, if you hit an object enough times, it will break and blow up into many small pieces. The pieces are cleaned up after a small delay, so no need to worry about objects flooding the scene.

The next commit will fix the colliders on the scaffold when it gets scaled to a different size. the mesh and colliders scale differently, causing invisible walls to appear.

* Fix Scaffolding

This commit finalizes my changes to the object system and fixes the scaffolding issue from earlier. The scaffold now uses a static mesh collider to calculate its colliders

---------

Co-authored-by: FarVillage117 <143549383+FarVillage117@users.noreply.github.com>