Add Extra Interior Design to Chambers
This commit adds some extra interior to the chambers we already have. This includes some pillars and chains with lights to break the boxed feeling our original chambers had.

Note: there are now many lights in the game due to changes I made to the lanterns. This is starting to cause a bit of a performance drop, but nothing too serious at the moment.