Create Basic Boss Room
This commit makes the basic boss room. There is much room for improvement, but time constraints have been taken into consideration. Next commits will focus on the following:

1) Performance improvements. The boss room takes up a lot of power to process because of many physics checks (30%) and rendering (15%). Recommendations for improvements are to make objects static whenever possible, remove rigidbodies from objects that dont need them, remove unnecessary colliders, and introduce a LOD system.

2) Boss fight. I created the basis for areas that the boss will fight the player. All that's needed to do is make the boss fight back against the player. I also intend to make some animations in Unity so the boss feels alive.

Side note: lanterns lost their bloom effect while in game view or prefab view because of the lack of post-processing. If we want this effect to continue, we need to introduce post-processing to the game.