Finalize Boss Chamber Look
This commit will finalize the boss chamber visuals for the meantime. If time permits, we may come back to add more to it, like lighting, trim, and wall decoration to break up the flat geometry. For now, this will suffice.

Next commits will focus on optimization, introducing the LOGGroup component to objects to control their level of detail from distances. I will also explore solutions to improve rigidbody performance for the chains, since they are the main culprit for performance issues when it comes to physics.