Add Level of Detail to Objects
This commit adds a level of detail to all of the objects we have used in the game so far. However, some values may need some adjustment. One caveat to reduced level of detail is distant objects appear more grainy.

Next commit will focus on reducing the physics cost of chains, since they are the main culprit for performance drops now.