Add Dynamic Kinematic System for Chain Rididbodies
This commit makes chains kinematic when the player is too far away from them to matter. Running the game shows that the physics update check went from taking ~35% of computing power down to ~10%. The majority of computing power is going to the editorloop, which is a good thing becuase that does not matter when building the game; it only affects the Unity editor.

While the physics update processing is lowered, it is worthwhile to explore making this behaviour generic to implement on other objects in the scene. 10% is still quite high compared to the other operations (rendering, etc.), and almost every object in the scene has a rigidbody attached to it. Finding solutions to increase kinematic states will helps imrpove performance at all stages within the game.