Begin Boss Fight Stuff (#46)
This commit begins adding stuff for the boss fight. I will focus on completing the boss fight for sprint 4, as I ran out of time.

I made a rig for the boss in Blender, and added a wiggle animation. This animation will be used when the boss chases the player.

The boss uses a navmeshagent to pathfind around the boss chamber. however, there are some things to fix up with the movement. The boss looks 90 degrees in the wrong direction, which will have to be fixed in sprint 4. We will also need to begin baking the nav mesh surfaces into our used chambers.