Jeremy roalef misc changes (#51)
* Update Shooting Behaviour

This commit updates where the player projectiles will shoot from. Before, the shooting position was determined by a transform position in the world. Now, it is based off of the main camera's position in the world. This ensures that shooting always happens from the center of the screen.

* Update Projectile Sticking Behavior

This commit updates how projectiles stick to surface. Before, the projectile would stick to along the normal of objects. This caused the projectile to snap its rotation away from the rotation it was shooting from.

* Create Pause System/Pause Manager

This commit adds a pause system to the game. Since the pause system will have to work with UI elements, I left the system in a barebones state until we can hook up the unpausing through UI elements,

Some important considerations for how the pausing works:

-Pausing the game means setting the Time.timescale variable to 0. This disables anything using Time.DeltaTime or any physics-based movement. However, scripts that use Update and LateUpdate will still be called. If there is any code related to the game in these Unity events that is not dependent on time scale, the pausing will not work as expected.

-Coroutines will interact with Time.timescale. However, if you are using WaitForSecondsRealtime, then the system will not be paused. Our code does not use this object, so our coroutines are currently safe.

* Final Touches to Misc Changes