Add Charged Shot Weapon
This commit adds the charged weapon shot, fully functioning. The weapon is currently equipped on the player and will have to be toggled on/off depending on the selected character.

I aimed for full control over the weapon's functionality. Every field that we would want to tinker with has been serialized. The only thing not configured is the connection to the augment system. Here's a list of lever we can pull:
-fire rate
-damage brackets
-base delay between shots
-random additional delay between shots
-total hit scanners (the things that charge the weapon)
-base max damage bonus
-perfect accuracy damage bonus