Add Shotgun W/ Shotgun Spread
This commit adds the final weapon into our game: the shotgun. The shotgun has a spread value that can be controller, though high spread values may result in unwanted results. Try keeping the spread beneath 5 degrees for accurate results.

There were issues with the UI crosshair. I tried to make it shrink/expand in regards to the spread of the shotgun, but I could not get an accurate crosshair