Object assets (#56)
* Added New FBX Files

* Move Files, Make Objects into Prefabs

* Mine Rooms Made

* FBX Hellscape Files Added

* Added New Door Files

* Changed Hallway Sizes

* Filled Hallway Prefabs

* Added 2 FBX Enemy Files

* Add Health Script & IDamageable Interface || Update Construction Site Prefabs

This commit adds some core systems to the game: health and damageable interface. These components can be used on anythign that has health behaviour or needs to take damage in the world.

A base object prefab was created, and I am going through the process of updating all of our original object prefabs to follow this new system.

Additionally, this commit introduces the ProBuilder package. This will be used to cut objects into small components. Then, we may make objects destructable.

* Update Mining Center Prefabs

This commit updates all of the mining center prefabs to follow the prefab variant structure.

All that is left to do now is replace all the original prefab references in the chambers and use the new prefab variants

* Update Object Prefab References in Chambers & Player

This commit finalizes the changes for all references to the original prefabs and makes them reference the new prefab variants.

* Add Hellscape Prefab Variants

This commit adds prefab variants to the Hellscape fbx files

* Update Object Colliders

This commit updates many of the colliders on the object prefabs. I used primitive colliders to make up the complex shapes. Additionally, I made the rock hook object with the chain act like a physical chain thanks to the hinge joint object component. We should explore adding more chains to our levels

* Add Destroyed Construction Site Objects

This commit adds the destroyed version of the objects for the construction site. Future commits will get the remaining objects in. First, I will test the code to ensure everything works correctly when destroyed the real game objects.

* Add Object Destruction

this commit adds object destruction. Now, if you hit an object enough times, it will break and blow up into many small pieces. The pieces are cleaned up after a small delay, so no need to worry about objects flooding the scene.

The next commit will fix the colliders on the scaffold when it gets scaled to a different size. the mesh and colliders scale differently, causing invisible walls to appear.

* Fix Scaffolding

This commit finalizes my changes to the object system and fixes the scaffolding issue from earlier. The scaffold now uses a static mesh collider to calculate its colliders

* Added New Enemies & Altar FBX Files

* Add Extra Interior Design to Chambers

This commit adds some extra interior to the chambers we already have. This includes some pillars and chains with lights to break the boxed feeling our original chambers had.

Note: there are now many lights in the game due to changes I made to the lanterns. This is starting to cause a bit of a performance drop, but nothing too serious at the moment.

* AddedBoardedWindowFBX

* Initial Start of Construction Site

This commit pushes the start of the construction site boss chamber. Next commit I push will finish the construction site. Then, I will work on the boss fight.

* Create Basic Boss Room

This commit makes the basic boss room. There is much room for improvement, but time constraints have been taken into consideration. Next commits will focus on the following:

1) Performance improvements. The boss room takes up a lot of power to process because of many physics checks (30%) and rendering (15%). Recommendations for improvements are to make objects static whenever possible, remove rigidbodies from objects that dont need them, remove unnecessary colliders, and introduce a LOD system.

2) Boss fight. I created the basis for areas that the boss will fight the player. All that's needed to do is make the boss fight back against the player. I also intend to make some animations in Unity so the boss feels alive.

Side note: lanterns lost their bloom effect while in game view or prefab view because of the lack of post-processing. If we want this effect to continue, we need to introduce post-processing to the game.

* Finalize Boss Chamber Look

This commit will finalize the boss chamber visuals for the meantime. If time permits, we may come back to add more to it, like lighting, trim, and wall decoration to break up the flat geometry. For now, this will suffice.

Next commits will focus on optimization, introducing the LOGGroup component to objects to control their level of detail from distances. I will also explore solutions to improve rigidbody performance for the chains, since they are the main culprit for performance issues when it comes to physics.

* Add Level of Detail to Objects

This commit adds a level of detail to all of the objects we have used in the game so far. However, some values may need some adjustment. One caveat to reduced level of detail is distant objects appear more grainy.

Next commit will focus on reducing the physics cost of chains, since they are the main culprit for performance drops now.

* Add Dynamic Kinematic System for Chain Rididbodies

This commit makes chains kinematic when the player is too far away from them to matter. Running the game shows that the physics update check went from taking ~35% of computing power down to ~10%. The majority of computing power is going to the editorloop, which is a good thing becuase that does not matter when building the game; it only affects the Unity editor.

While the physics update processing is lowered, it is worthwhile to explore making this behaviour generic to implement on other objects in the scene. 10% is still quite high compared to the other operations (rendering, etc.), and almost every object in the scene has a rigidbody attached to it. Finding solutions to increase kinematic states will helps imrpove performance at all stages within the game.

* Augment Icons FBX Files Added

* More FBX Augment Icons

* update metadata

* Remove Duplicate Game Settings Script

* 3x3 Chamber Decorated

* 3x1 Chamber Decorated

---------

Co-authored-by: FarVillage117 <143549383+FarVillage117@users.noreply.github.com>